+++ locksmith/src/models/index.ts
@@ -3,7 +3,6 @@ import { Verifier } from './verifier'
 import { User } from './user'
 import { UserReference } from './userReference'
 import { Lock } from './lock'
-import { Transaction } from './transaction'
 import { AuthorizedLock } from './authorizedLock'
 import { LockMetadata } from './lockMetadata'
 import { KeyMetadata } from './keyMetadata'
@@ -48,7 +47,6 @@ sequelize.addModels([
   RefreshToken,
   StripeConnectLock,
   StripeCustomer,
-  Transaction,
   User,
   UserReference,
   UserTokenMetadata,
@@ -59,14 +57,12 @@ sequelize.addModels([
 
 User.removeAttribute('id')
 Lock.removeAttribute('id')
-Transaction.removeAttribute('id')
 LockMetadata.removeAttribute('id')
 StripeCustomer.removeAttribute('id')
 
 export * from './user'
 export * from './userReference'
 export * from './lock'
-export * from './transaction'
 export * from './authorizedLock'
 export * from './usertokenmetadata'
 export * from './stripeCustomer'