+++ smart-contracts/test/Lock/maxNumberOfKeys.js
@@ -11,11 +11,11 @@ const {
 
 const keyPrice = ethers.utils.parseEther('0.01')
 
-contract('Lock / setMaxNumberOfKeys', () => {
+contract('Lock / maxNumberOfKeys', () => {
   let unlock
   let lock
 
-  describe('update the number of keys available in a lock', () => {
+  describe('prevent from buying more keys than defined supply', () => {
     beforeEach(async () => {
       const { unlockEthers: unlockDeployed } = await deployContracts()
       unlock = unlockDeployed
@@ -36,7 +36,7 @@ contract('Lock / setMaxNumberOfKeys', () => {
       lock = PublicLock.attach(newLockAddress)
     })
 
-    it('should increase max number of keys', async () => {
+    it('should prevent from buying more keys from defined', async () => {
       const [, ...buyers] = await ethers.getSigners()
 
       // buy 10 key
@@ -46,38 +46,17 @@ contract('Lock / setMaxNumberOfKeys', () => {
       await reverts(purchaseKey(lock, buyers[11].address), 'LOCK_SOLD_OUT')
 
       // increase supply
-      await lock.setMaxNumberOfKeys(12)
+      await lock.updateLockConfig(
+        await lock.expirationDuration(),
+        12,
+        await lock.maxKeysPerAddress()
+      )
 
       // actually buy the key
       const { to } = await purchaseKey(lock, buyers[11].address)
 
       assert.equal(to, buyers[11].address)
       assert.equal(await lock.maxNumberOfKeys(), 12)
-    })
-
-    it('should prevent from setting a value lower than total supply', async () => {
-      // buy 10 keys
-      await purchaseKeys(lock, 10)
-
-      // increase supply
-      await reverts(lock.setMaxNumberOfKeys(5), 'SMALLER_THAN_SUPPLY')
-    })
-
-    it('should allow setting a value equal to current total supply', async () => {
-      // buy 10 keys
-      const [, ...buyers] = await ethers.getSigners()
-      await purchaseKeys(lock, 10)
-
-      // set max keys to total supply
-      const totalSupply = await lock.totalSupply()
-      await lock.setMaxNumberOfKeys(totalSupply)
-      assert.equal(
-        (await lock.maxNumberOfKeys()).toString(),
-        totalSupply.toString()
-      )
-
-      // try to buy another key exceding totalSupply
-      await reverts(purchaseKey(lock, buyers[11].address), 'LOCK_SOLD_OUT')
     })
   })
 })