+++ smart-contracts/test/Lock/onKeyTransferHook.js
@@ -21,11 +21,10 @@ contract('Lock / onKeyTransfer hook', (accounts) => {
       ADDRESS_ZERO,
       ADDRESS_ZERO,
       testEventHooks.address,
+      ADDRESS_ZERO,
       ADDRESS_ZERO
     )
     keyPrice = await lock.keyPrice()
-
-    await lock.setMaxKeysPerAddress(10)
   })
 
   beforeEach(async () => {
@@ -83,6 +82,7 @@ contract('Lock / onKeyTransfer hook', (accounts) => {
         ADDRESS_ZERO,
         ADDRESS_ZERO,
         accounts[1],
+        ADDRESS_ZERO,
         ADDRESS_ZERO
       ),
       'INVALID_HOOK(4)'