+++ smart-contracts/test/Lock/owners.js
@@ -8,7 +8,6 @@ contract('Lock / owners', (accounts) => {
   const keyOwners = accounts.slice(1, 6)
   before(async () => {
     lock = await deployLock()
-    await lock.setMaxKeysPerAddress(10)
     await lock.updateTransferFee(0) // disable the transfer fee for this test
   })